LayoutTests/sputnik/Conformance/12_Statement/12.11_The_switch_Statement/S12.11_A3_T5.html

<html>
<head>
<meta charset='utf-8'>
<style>
.pass {
    font-weight: bold;
    color: green;
}
.fail {
    font-weight: bold;
    color: red;
}
</style>

<script>
if (window.testRunner)
    testRunner.dumpAsText();

function SputnikError(message)
{
    this.message = message;
}

SputnikError.prototype.toString = function ()
{
    return 'SputnikError: ' + this.message;
};

var sputnikException;

function testPrint(msg)
{
    var span = document.createElement("span");
    document.getElementById("console").appendChild(span); // insert it first so XHTML knows the namespace 
    span.innerHTML = msg + '<br />';
}

function escapeHTML(text)
{
    return text.toString().replace(/&/g, "&amp;").replace(/</g, "&lt;");
}

function printTestPassed(msg)
{
    testPrint('<span><span class="pass">PASS</span> ' + escapeHTML(msg) + '</span>');
}

function printTestFailed(msg)
{
    testPrint('<span><span class="fail">FAIL</span> ' + escapeHTML(msg) + '</span>');
}

function testFailed(msg)
{
    throw new SputnikError(msg);
}

var successfullyParsed = false;
</script>

</head>
<body>
<p>S12.11_A3_T5</p>
<div id='console'></div>
<script>
try {

/**
 * @name: S12.11_A3_T5;
 * @section: 12.11;
 * @assertion: Syntax constructions of switch statement;
 * @description: Introducing statement not followed by "case" keyword;
 * @negative;
*/

function SwitchTest(value){
  var result = 0;
  
  switch(value) {
  	result =2;
    case 0:
      result += 2;
    default:
      result += 32;
      break;
  }
  
  return result;
}

var x = SwitchTest(0);

} catch (ex) {
    sputnikException = ex;
}

var successfullyParsed = true;
</script>

<script>
if (!successfullyParsed)
    printTestPassed('Expected parsing failure');
else if (sputnikException)
    printTestPassed(sputnikException);
else
    printTestFailed("No error detected");
testPrint('<br /><span class="pass">TEST COMPLETE</span>');
</script>
</body>
</html>
